refactor(usage): move the breakdown's data into a hook
The section's queries, dimension state and row derivation were in the
component. They are now in useUsageBreakdown, leaving 42 lines that wire the
hook to the view.

The split existed to separate stateful logic from rendering, which is what a
hook does without needing a second component to hold it. ScopeTotal still
renders, because project and environment need one request per key and a hook
cannot mount a query per item of a list that changes length.

Also drops a guard that could not fire: with no project selected the
environments query is skipped, so scopes is already empty.
